+++ src/components/color-picker-swatch/color-picker-swatch.e2e.ts
@@ -18,13 +18,14 @@ describe("calcite-color-picker-swatch", () => {
     accessible(`<calcite-color-picker-swatch active color='#c0ffee'></calcite-color-picker-swatch>`);
   });
 
-  it("has defaults", () =>
+  describe("defaults", () => {
     defaults("calcite-color-picker-swatch", [
       {
         propertyName: "active",
         defaultValue: false
       }
-    ]));
+    ]);
+  });
 
   describe("reflects", () => {
     reflects("calcite-color-picker-swatch", [