+++ src/components/color-picker/color-picker.e2e.ts
@@ -59,7 +59,7 @@ describe("calcite-color-picker", () => {
     renders("calcite-color-picker", { display: "inline-block" });
   });
 
-  it("has defaults", async () =>
+  describe("defaults", () => {
     defaults("calcite-color-picker", [
       {
         propertyName: "allowEmpty",
@@ -105,7 +105,8 @@ describe("calcite-color-picker", () => {
         propertyName: "value",
         defaultValue: "#007ac2"
       }
-    ]));
+    ]);
+  });
 
   // #408047 is a color in the middle of the color field
   it("can be disabled", () => disabled("<calcite-color-picker value='#408047'></calcite-color-picker>"));
@@ -649,6 +650,34 @@ describe("calcite-color-picker", () => {
     expect(afterDragHsv.h).toBe(beforeDragHsv.h);
     expect(afterDragHsv.s).toBeGreaterThan(beforeDragHsv.s);
     expect(afterDragHsv.v).toBe(beforeDragHsv.v);
+  });
+
+  it("does not wrap the hue slider thumb when dragging past the edge", async () => {
+    const page = await newE2EPage();
+    await page.setContent(`<calcite-color-picker></calcite-color-picker>`);
+    const [hueSliderX] = await getElementXY(page, "calcite-color-picker", `.${CSS.hueSlider}`);
+
+    let [hueScopeX, hueScopeY] = await getElementXY(page, "calcite-color-picker", `.${CSS.hueScope}`);
+
+    await page.mouse.move(hueScopeX, hueScopeY);
+    await page.mouse.down();
+    await page.mouse.move(0, hueScopeY);
+    await page.mouse.up();
+    await page.waitForChanges();
+
+    [hueScopeX, hueScopeY] = await getElementXY(page, "calcite-color-picker", `.${CSS.hueScope}`);
+
+    expect(hueScopeX).toBe(hueSliderX);
+
+    await page.mouse.move(hueScopeX, hueScopeY);
+    await page.mouse.down();
+    await page.mouse.move(hueScopeX + DIMENSIONS.m.slider.width, hueScopeY);
+    await page.mouse.up();
+    await page.waitForChanges();
+
+    [hueScopeX] = await getElementXY(page, "calcite-color-picker", `.${CSS.hueScope}`);
+
+    expect(hueScopeX).toBe(hueSliderX + DIMENSIONS.m.slider.width - 1);
   });
 
   describe("unsupported value handling", () => {