+++ src/components/color-picker/color-picker.tsx
@@ -617,7 +617,7 @@ export class ColorPicker
       } else if (clientX < bounds.x) {
         samplingX = 0;
       } else {
-        samplingX = bounds.width;
+        samplingX = bounds.width - 1;
       }
 
       if (clientY < bounds.y + bounds.height && clientY > bounds.y) {