+++ src/components/filter/filter.e2e.ts
@@ -35,7 +35,7 @@ describe("calcite-filter", () => {
     ]);
   });
 
-  it("has defaults", async () =>
+  describe("defaults", () => {
     defaults("calcite-filter", [
       {
         propertyName: "disabled",
@@ -49,7 +49,8 @@ describe("calcite-filter", () => {
         propertyName: "scale",
         defaultValue: "m"
       }
-    ]));
+    ]);
+  });
 
   it("sets scale on the input", async () => {
     const scale = "s";