+++ src/components/input-text/input-text.e2e.ts
@@ -41,7 +41,7 @@ describe("calcite-input-text", () => {
     ]);
   });
 
-  it("has defaults", async () =>
+  describe("defaults", () => {
     defaults("calcite-input-text", [
       {
         propertyName: "status",
@@ -59,7 +59,8 @@ describe("calcite-input-text", () => {
         propertyName: "value",
         defaultValue: ""
       }
-    ]));
+    ]);
+  });
 
   it("can be disabled", () => disabled("calcite-input-text"));