+++ src/components/input-time-picker/input-time-picker.e2e.ts
@@ -44,12 +44,13 @@ describe("calcite-input-time-picker", () => {
 
   it.skip("supports t9n", () => t9n("calcite-input-time-picker"));
 
-  it("has defaults", async () =>
+  describe("defaults", () => {
     defaults("calcite-input-time-picker", [
       { propertyName: "scale", defaultValue: "m" },
       { propertyName: "step", defaultValue: 60 },
       { propertyName: "overlayPositioning", defaultValue: "absolute" }
-    ]));
+    ]);
+  });
 
   describe("reflects", () => {
     reflects(`calcite-input-time-picker`, [