+++ src/components/input/input.e2e.ts
@@ -21,6 +21,7 @@ describe("calcite-input", () => {
 
   /**
    * This helper wraps number typing to work around test instability
+   *
    * @param page
    * @param numberAsText
    */
@@ -63,7 +64,7 @@ describe("calcite-input", () => {
     ]);
   });
 
-  it("has defaults", async () =>
+  describe("defaults", () => {
     defaults("calcite-input", [
       {
         propertyName: "status",
@@ -89,7 +90,8 @@ describe("calcite-input", () => {
         propertyName: "value",
         defaultValue: ""
       }
-    ]));
+    ]);
+  });
 
   it("can be disabled", () => disabled("calcite-input"));