+++ src/components/navigation/navigation.e2e.ts
@@ -20,13 +20,14 @@ describe("calcite-navigation", () => {
     ]);
   });
 
-  it("defaults", () =>
+  describe("defaults", () => {
     defaults("calcite-navigation", [
       {
         propertyName: "navigationAction",
         defaultValue: false
       }
-    ]));
+    ]);
+  });
 
   describe("is focusable", () => {
     focusable(html`<calcite-navigation navigation-action></calcite-navigation>`, {