+++ src/tests/commonTests.ts
@@ -8,6 +8,7 @@ import { JSX } from "../components";
 import { hiddenFormInputSlotName } from "../utils/form";
 import { MessageBundle } from "../utils/t9n";
 import { GlobalTestProps, skipAnimations } from "./utils";
+import { InteractiveHTMLElement } from "../utils/interactive";
 
 expect.extend(toHaveNoViolations);
 
@@ -52,8 +53,9 @@ async function simplePageSetup(componentTagOrHTML: TagOrHTML): Promise<E2EPage> 
  *
  * Note that this helper should be used within a describe block.
  *
+ * @example
  * describe("accessible"), () => {
- * accessible(`<calcite-tree></calcite-tree>`);
+ *    accessible(`<calcite-tree></calcite-tree>`);
  * });
  * @param {ComponentTestSetup} componentTestSetup - A component tag, html, or the tag and e2e page for setting up a test
  */
@@ -73,8 +75,9 @@ export function accessible(componentTestSetup: ComponentTestSetup): void {
 /**
  * Note that this helper should be used within a describe block.
  *
+ * @example
  * describe("renders", () => {
- * renders(`<calcite-tree></calcite-tree>`);
+ *    renders(`<calcite-tree></calcite-tree>`);
  * });
  * @param {string} componentTagOrHTML - the component tag or HTML markup to test against
  * @param {object} options - additional options to assert
@@ -117,6 +120,7 @@ export async function renders(
  * }
  * ])
  * })
+ *
  * @param {string} componentTagOrHTML - the component tag or HTML markup to test against
  * @param {object[]} propsToTest - the properties to test
  * @param {string} propsToTest.propertyName - the property name
@@ -170,26 +174,45 @@ function propToAttr(name: string): string {
 
 /**
  * Helper for asserting that a property's value is its default
+ *
+ * Note that this helper should be used within a describe block.
+ *
+ * @example
+ * describe("defaults", () => {
+ *    defaults("calcite-action", [
+ *      {
+ *        propertyName: "active",
+ *        defaultValue: false
+ *      },
+ *      {
+ *        propertyName: "appearance",
+ *        defaultValue: "solid"
+ *      }
+ *    ])
+ * })
+ *
  * @param {string} componentTagOrHTML - the component tag or HTML markup to test against
  * @param {object[]} propsToTest - the properties to test
  * @param {string} propsToTest.propertyName - the property name
  * @param {any} propsToTest.value - the property value
  */
-export async function defaults(
+export function defaults(
   componentTagOrHTML: TagOrHTML,
   propsToTest: {
     propertyName: string;
     defaultValue: any;
   }[]
-): Promise<void> {
-  const page = await simplePageSetup(componentTagOrHTML);
-  const element = await page.find(getTag(componentTagOrHTML));
+): void {
+  it("has property defaults", async () => {
+    const page = await simplePageSetup(componentTagOrHTML);
+    const element = await page.find(getTag(componentTagOrHTML));
 
-  for (const propAndValue of propsToTest) {
-    const { propertyName, defaultValue } = propAndValue;
-    const prop = await element.getProperty(propertyName);
-    expect(prop).toEqual(defaultValue);
-  }
+    for (const propAndValue of propsToTest) {
+      const { propertyName, defaultValue } = propAndValue;
+      const prop = await element.getProperty(propertyName);
+      expect(prop).toEqual(defaultValue);
+    }
+  });
 }
 
 /**
@@ -197,6 +220,7 @@ export async function defaults(
  *
  * Note that this helper should be used within a describe block.
  *
+ * @example
  * describe("honors hidden attribute", () => {
  * hidden("calcite-accordion")
  * });
@@ -217,6 +241,7 @@ export async function hidden(componentTagOrHTML: TagOrHTML): Promise<void> {
 interface FocusableOptions {
   /**
    * use this to pass an ID to setFocus()
+   *
    * @deprecated components should no longer use a focusId parameter for setFocus()
    */
   focusId?: string;
@@ -237,6 +262,7 @@ interface FocusableOptions {
  *
  * Note that this helper should be used within a describe block.
  *
+ * @example
  * describe("is focusable", () => {
  * focusable(`calcite-input-number`, { shadowFocusTargetSelector: "input" })
  * });
@@ -273,6 +299,7 @@ export function focusable(componentTagOrHTML: TagOrHTML, options?: FocusableOpti
 
 /**
  * Helper for asserting slots.
+ *
  * @param {string} componentTagOrHTML - the component tag or HTML markup to test against
  * @param slots - a component's SLOTS resource object or an array of slot names
  * @param includeDefaultSlot - when true, it will run assertions on the default slot
@@ -396,6 +423,7 @@ interface LabelableOptions extends Pick<FocusableOptions, "focusTargetSelector" 
 
 /**
  * Helper for asserting label clicking functionality works.
+ *
  * @param {string} componentTagOrHtml - the component tag or HTML used to test label support
  * @param {LabelableOptions} [options] - labelable options
  */
@@ -545,8 +573,9 @@ interface FormAssociatedOptions {
  *
  * Note that this helper should be used within a describe block.
  *
+ * @example
  * describe("form-associated), () => {
- * formAssociated("calcite-component", { testValue: 1337 });
+ *    formAssociated("calcite-component", { testValue: 1337 });
  * });
  * @param {string} componentTagOrHtml - the component tag or HTML markup to test against
  * @param {FormAssociatedOptions} options - form associated options
@@ -852,6 +881,7 @@ async function getTagAndPage(componentTestSetup: ComponentTestSetup): Promise<Ta
 
 /**
  * Helper to test the disabled prop disabling user interaction.
+ *
  * @param {ComponentTestSetup} componentTestSetup - A component tag, html, or the tag and e2e page for setting up a test
  * @param {DisabledOptions} [options={ focusTarget: "host" }] - disabled options
  */
@@ -998,11 +1028,35 @@ export async function disabled(
       expect(spy).toHaveReceivedEventTimes(eventsExpectedToBubble.includes(spy.eventName) ? 2 : 1);
     }
   });
+
+  // this needs to run in the browser context to ensure disabling and events fire immediately after being set
+  await page.$eval(
+    tag,
+    (component: InteractiveHTMLElement, allExpectedEvents: string[]) => {
+      component.disabled = false;
+      allExpectedEvents.forEach((event) => component.dispatchEvent(new MouseEvent(event)));
+
+      component.disabled = true;
+      allExpectedEvents.forEach((event) => component.dispatchEvent(new MouseEvent(event)));
+    },
+    allExpectedEvents
+  );
+
+  assertOnMouseAndPointerEvents(eventSpies, (spy) => {
+    if (spy.eventName === "click") {
+      // some components emit more than one click event (e.g., from calling `click()`),
+      // so we check if at least one event is received
+      expect(spy.length).toBeGreaterThanOrEqual(3);
+    } else {
+      expect(spy).toHaveReceivedEventTimes(eventsExpectedToBubble.includes(spy.eventName) ? 4 : 2);
+    }
+  });
 }
 
 /**
  * This helper will test if a floating-ui-owning component has configured the floating-ui correctly.
  * At the moment, this only tests if the scroll event listeners are only active when the floating-ui is displayed.
+ *
  * @param componentTagOrHTML - the component tag or HTML markup to test against
  * @param togglePropName - the component property that toggles the floating-ui
  * @param options - the floating-ui owner test configuration
@@ -1084,6 +1138,7 @@ export async function floatingUIOwner(
 
 /**
  * Helper to test t9n component setup
+ *
  * @param {ComponentTestSetup} componentTestSetup - A component tag, html, or the tag and e2e page for setting up a test
  */
 export async function t9n(componentTestSetup: ComponentTestSetup): Promise<void> {